# server/openapi.yaml
openapi: 3.1.0
info:
  title: UCT ERP Secured API Portal
  description: |
    Production-grade, enterprise-ready API endpoints for Universidade Católica Timorense (UCT) ERP.
    
    ## Authentication
    All endpoints (except `/api/auth/login`) require authentication via JWT Bearer Token.
    Provide the token in the headers as:
    `Authorization: Bearer <JWT_TOKEN>`

    ## Role-Based Access Control (RBAC)
    Operations and visibility are restricted by portal role:
    - **Administrator**: Unrestricted read/write/delete access.
    - **Student**: Access restricted to own student details, attendance, and fee transactions.
    - **Staff/Faculty**: Access to manage students (read/write) and mark attendance.
    - **HR**: Access to manage employee records and payroll.
    - **Accounts**: Access to manage transactions, ledgers, and fee structures.
    - **Transport**: Access to manage transport routes, vehicles, and allocations.
  version: 2.0.0
servers:
  - url: http://localhost:5000
    description: Local Express Backend Server
paths:
  /api/auth/login:
    post:
      summary: User Portal Sign In
      description: Validates user credentials and portal role selection, returning a signed JWT token.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - role
              properties:
                username:
                  type: string
                  description: Login username or email address
                  example: admin
                password:
                  type: string
                  description: Account password
                  example: Admin@UctErp2026
                role:
                  type: string
                  enum: [Administrator, Student, Staff/Faculty, HR, Accounts, Transport, Agent]
                  description: Chosen portal login role
                  example: Administrator
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  token:
                    type: string
                    description: JWT Bearer authorization token
                  user:
                    type: object
                    properties:
                      username:
                        type: string
                      role:
                        type: string
        '400':
          description: Missing required fields
        '401':
          description: Invalid credentials or incorrect portal role
  /api/auth/change-password:
    post:
      summary: Change Account Password
      description: Updates the logged-in user's password hash.
      tags:
        - Authentication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - currentPassword
                - newPassword
              properties:
                username:
                  type: string
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Password updated successfully
        '401':
          description: Incorrect current password
        '404':
          description: User not found
...
  /api/fees:
    get:
      summary: Get Fee Transactions
      description: |
        Lists receipts/ledger transactions.
        - **Student role**: Returns only fee receipts matching their own student ID.
        - **Accounts, Administrator**: Returns all fee receipts.
        - **Other roles**: Returns 403 Forbidden.
      tags:
        - Accounts & Fees
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of receipts
        '403':
          description: Insufficient permissions

    post:
      summary: Collect Fee / Create Receipt
      description: Registers a new payment transaction and decreases the student's due fee balance (Restricted to Administrator and Accounts).
      tags:
        - Accounts & Fees
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - student_id
                - amount
                - payment_mode
              properties:
                student_id:
                  type: integer
                amount:
                  type: number
                payment_mode:
                  type: string
                date:
                  type: string
                head_wise_details:
                  type: object
      responses:
        '201':
          description: Transaction created and balance adjusted successfully
        '403':
          description: Insufficient permissions
  /api/generic/{module_key}/{record_id}:
    parameters:
      - name: module_key
        in: path
        required: true
        schema:
          type: string
          description: Target module key (e.g. transport, hostel, student-attendance)
      - name: record_id
        in: path
        required: true
        schema:
          type: string
          description: Document record ID (e.g. routes, allocations, records)
    get:
      summary: Get Generic Module Config/Record
      description: |
        Fetches JSON record for secondary modules under role rules:
        - **Transport role**: Allowed to read/write `transport` key.
        - **Accounts role**: Allowed to read/write `accounts` and `fees` keys.
        - **Staff/Faculty role**: Allowed to read/write `student-attendance`, `timetable`, `lesson-plan` keys, and read `hostel`, `transport` keys.
        - **Student role**: Allowed to read keys `hostel` (filtered to their details), `student-attendance` (filtered to their attendance), `timetable`, `lesson-plan`, `certificate`, `document`.
      tags:
        - Secondary Modules
      security:
        - bearerAuth: []
      responses:
        '200':
          description: JSON object or array stored under requested key
        '403':
          description: Insufficient permissions

    post:
      summary: Save Generic Module Config/Record
      description: Overwrites or updates a generic key-value JSON entry (Restricted by module key write-rules).
      tags:
        - Secondary Modules
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Entry saved/upserted successfully
        '403':
          description: Insufficient permissions
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Standard Authorization Bearer token header.
...